Remove notification for specific EDI case.
The ONLY case in which it is possible to create a need for a
notification which would not be caught elsewhere in the process is via
CV upload in the gluedb interface.

In every other case you would be executing a 'double notify' and getting
to notices.

As the CV upload handling is out of scope of this, and given the high
chance for double-notification, I have removed notification code from
EDI import.